Add rubygem-local_time 1.0.3
Local Time is a Rails engine with helpers and JavaScript for displaying times
and dates to users in their local time. The helpers render a <time> element and
the JavaScript swoops in to convert and format. The helper output is ideal for
caching since it's always in UTC time.

WWW: https://github.com/basecamp/local_time